Move browser checks to browser.js
