+++ core/rfb.js
@@ -12,7 +12,7 @@
 
 import * as Log from './util/logging.js';
 import { decodeUTF8 } from './util/strings.js';
-import { browserSupportsCursorURIs, isTouchDevice } from './util/browsers.js';
+import { supportsCursorURIs, isTouchDevice } from './util/browser.js';
 import EventTargetMixin from './util/eventtarget.js';
 import Display from "./display.js";
 import Keyboard from "./input/keyboard.js";
@@ -1275,7 +1275,7 @@ RFB.prototype = {
         encs.push(encodings.pseudoEncodingFence);
         encs.push(encodings.pseudoEncodingContinuousUpdates);
 
-        if (browserSupportsCursorURIs() &&
+        if (supportsCursorURIs() &&
             !isTouchDevice && this._fb_depth == 24) {
             encs.push(encodings.pseudoEncodingCursor);
         }
