+++ tests/test.keyboard.js
@@ -4,13 +4,7 @@ var expect = chai.expect;
 import sinon from '../vendor/sinon.js';
 
 import Keyboard from '../core/input/keyboard.js';
-
-function isIE() {
-    return navigator && !!(/trident/i).exec(navigator.userAgent);
-}
-function isEdge() {
-    return navigator && !!(/edge/i).exec(navigator.userAgent);
-}
+import * as browser from '../core/util/browser.js';
 
 /* jshint newcap: false, expr: true */
 describe('Key Event Handling', function() {
@@ -30,7 +24,7 @@ describe('Key Event Handling', function() {
 
     describe('Decode Keyboard Events', function() {
         it('should decode keydown events', function(done) {
-            if (isIE() || isEdge()) this.skip();
+            if (browser.isIE() || browser.isEdge()) this.skip();
             var kbd = new Keyboard(document);
             kbd.onkeyevent = function(keysym, code, down) {
                 expect(keysym).to.be.equal(0x61);
@@ -41,7 +35,7 @@ describe('Key Event Handling', function() {
             kbd._handleKeyDown(keyevent('keydown', {code: 'KeyA', key: 'a'}));
         });
         it('should decode keyup events', function(done) {
-            if (isIE() || isEdge()) this.skip();
+            if (browser.isIE() || browser.isEdge()) this.skip();
             var calls = 0;
             var kbd = new Keyboard(document);
             kbd.onkeyevent = function(keysym, code, down) {
@@ -136,7 +130,7 @@ describe('Key Event Handling', function() {
 
         describe('suppress the right events at the right time', function() {
             beforeEach(function () {
-                if (isIE() || isEdge()) this.skip();
+                if (browser.isIE() || browser.isEdge()) this.skip();
             });
             it('should suppress anything with a valid key', function() {
                 var kbd = new Keyboard(document, {});
@@ -166,7 +160,7 @@ describe('Key Event Handling', function() {
 
     describe('Fake keyup', function() {
         it('should fake keyup events for virtual keyboards', function(done) {
-            if (isIE() || isEdge()) this.skip();
+            if (browser.isIE() || browser.isEdge()) this.skip();
             var count = 0;
             var kbd = new Keyboard(document);
             kbd.onkeyevent = function(keysym, code, down) {
@@ -213,7 +207,7 @@ describe('Key Event Handling', function() {
             });
 
             it('should fake keyup events on iOS', function(done) {
-                if (isIE() || isEdge()) this.skip();
+                if (browser.isIE() || browser.isEdge()) this.skip();
                 var count = 0;
                 var kbd = new Keyboard(document);
                 kbd.onkeyevent = function(keysym, code, down) {
@@ -237,7 +231,7 @@ describe('Key Event Handling', function() {
 
     describe('Track Key State', function() {
         beforeEach(function () {
-            if (isIE() || isEdge()) this.skip();
+            if (browser.isIE() || browser.isEdge()) this.skip();
         });
         it('should send release using the same keysym as the press', function(done) {
             var kbd = new Keyboard(document);
